requirements.txt
onnx==1.13
onnxoptimizer==0.3.13
onnxruntime==1.18
matplotlib==3.5.1
numpy==1.23.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
openpyxl==3.1.2
pandas==1.5.3
pillow==9.3.0
plotly==5.14.1
scikit-learn==1.3.0
tensorboardx==2.6.2.2
timm==0.9.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchinfo==1.7.2
torchvision==0.15.2+cu118
transformers==4.28.1
